Take an int instead of a double. From a patch by jody@ximian.com.
2002-04-26  Christopher James Lahey  <clahey@ximian.com>

	* e-table-field-chooser-item.c (etfci_maybe_start_drag): Take an
	int instead of a double.  From a patch by jody@ximian.com.

svn path=/trunk/; revision=16606

+++ widgets/table/e-table-field-chooser-item.c
@@ -503,7 +503,7 @@ etfci_point (GnomeCanvasItem *item, double x, double y, int cx, int cy,
 }
 
 static gboolean
-etfci_maybe_start_drag (ETableFieldChooserItem *etfci, double x, double y)
+etfci_maybe_start_drag (ETableFieldChooserItem *etfci, int x, int y)
 {
 	if (!etfci->maybe_drag)
 		return FALSE;
